+++ app/src/main/java/com/example/personalphysicaltracker/database/TrackingRepository.kt
@@ -34,7 +34,7 @@ class TrackingRepository(app: Application) {
         }
     }
 
-    fun getTotalDurationByActivityType(activityType: String): Long {
+    fun getTotalDurationByActivityType(activityType: String): Double {
         return trackingDao.getTotalDurationByActivityType(activityType)
     }
